Fixed bug and added test for INT16 support
I only partially / incorrectly added INT16 support in a previous commit.
This fixes it.

+++ cpp/src/schema.cpp
@@ -66,7 +66,7 @@ StreamSchema StreamSchema::FromJson(std::string json_str) {
       } else if (type_str == "FLOAT") {
         type = FieldDefinition::FLOAT;
       } else if (type_str == "INT16") {
-        type = FieldDefinition::INT32;
+        type = FieldDefinition::INT16;
       } else if (type_str == "INT32") {
         type = FieldDefinition::INT32;
       } else if (type_str == "INT64") {
